KAN-103 refactor: (wip) code structure > index pannel
